requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages

# Data processing
pandas==2.2.0
requests==2.31.0

# Database
psycopg2-binary==2.9.9

# Data versioning
dvc==3.64.0

# Git operations
GitPython==3.1.45
